services/accounting/pyproject.toml
[tool.poetry]
name = "accounting"
version = "0.1.0"
description = ""
authors = ["Wojtek Krzywiec <wojciechkrzywiec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "~=1.2"
plotly = "~=4.14"
numpy = "~=1.20"
pyyaml = "*"
python-dateutil = "*"
pycountry = "^22.3.5"

[tool.poetry.group.dev.dependencies]
pytest = "~=6.2"
pytest-cov = "~=2.11"
pytest-sugar = "~=0.9.4"
pytest-mock = "~=3.5.1"
flake8 = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
